Refactor server routes to use searchParams
