+++ src/lib/utils.ts
@@ -1,8 +1,13 @@
 import { type ClassValue, clsx } from 'clsx'
 import { twMerge } from 'tailwind-merge'
 
-const environment = process.env.VERCEL_ENV
-
 export function cn(...inputs: ClassValue[]) {
   return twMerge(clsx(inputs))
+}
+
+export const baseUrl = process.env.BASE_URL
+
+// validate email link
+export const createValidateEmailLink = (token: string) => {
+  return `${baseUrl}/api/validate?token=${token}`
 }
